fixed the inline window's last char's visibility
When the styling was "min-content", we've noticed that the chars don't always positioned within their cell's border. This sometimes cause the last char of the row to be cut.
By changing the style to "fit-content", each char is now positioned within its cell's border, and as a result - the last char is fully visible.